Add support for all maps in Splinter Cell 1
Support for autosplitting added for all maps. The first map ever loaded -- 0_0_2_training -- shouldn't matter because timing starts just after the map is loaded.

note:
I'm not super proficient in C#, but I found the list of maps in this file and figured you just add to it to add more maps.